@@ -1,8 +1,8 @@
 package com.kayushi07.getablessing;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 
 /**
  * Created by Ayushi on 21-04-2018.
