@@ -11,7 +11,7 @@ import java.util.List;
 public class BlessedData {
 
     public static String[] getData() {
-        String[] blessings = {"May success be with you, always…Wishing you good luck!", "May you do well and come out with flying colors…ALL THE BEST!", "May you fly high in life & success be with you always.", "Try a little harder to be a little better. All the best!", "The future is bright and beautiful. Love it, strive for it and work for it!", "My heart is filled with joy at the sound of your good news. Best of luck to you dear friend.", "Be strong as you fight your way to success, the difficulty you face is nothing compared to the joy of success.", "Good luck to you as you move toward the next chapter life brings your way.", "Good luck and tons of best wishes. God bless you in whatever you do. This is my heartiest wish just for you.", "Here’s wishing you SUCCESS in everything you do! Good luck!", "Success comes only to those who believe in themselves and are prepared to win. Good Luck.", "I wish nothing but the best for you.", "Moving on to a new stage in life can be a challenging process. We wish you the best of luck in all of your future endeavors, you will be great!", "It’s kind of fun to do the impossible. All the best.", "Good LUCK to you as you move onto the next stage in your life. Without a doubt, you will continue to have SUCCESS in all areas!", "And here’s wishing you the very best for all the new ventures, that life has in store for you.", "I find my heart singing a chorus at your happy news. Good luck to you during this joyous time.", "Keep your best wishes, close to your heart and watch what happens.", "Don’t be afraid of moving forward. The future is bright.", "If you can dream it, you can do it. Best wishes for your bright future.", "I wish you all the success, happiness, and joy in life.", "You have worked incredibly hard. Good luck! We are proud of you.", "Don’t look back! For a bright future is ahead of you.", "You have proven yourself to be a very skilled individual who has the capacity to do great things with their life. Continue to make us proud as you face new challenges and adventures!", "Dream Big..Sparkle More..Shine Bright.", "Hope your burdens are feeling lighter, and each day is a little bit brighter.", "I am thrilled to know that luck has finally found you. Warmest wishes my friend.", "Sending you thoughtful wishes in all life’s endeavors. Good luck to you now and always.", "Life is what you make of it and I am happy to know that yours is filled with such good luck.", "Wishing you all the best in your future endeavors.", "Winning is not everything but wanting to win is everything Go Ahead…BEST OF LUCK!", "Here’s to wishing you the best of luck in all life has to offer you.", "Go confidently in the direction of your dreams. Live the life you have imagined. – Henry David Thoreau", "Do not go where the path may lead; go instead where there is no path and leave a trail. – Ralph Waldo Emerson", "Hurdles and difficulties are part of the life. Make that part your life a past and think about tomorrow, because what matters at the end of the day is tomorrow.", "Always keep your head up with a smile on the face. Giving best wishes and all the happiness in this world.", "May luck always follow you in whatever you do. May you achieve success as you walk through.", "Changing the face can change nothing, but facing the change can change everything. Good luck.", "Best wishes for you when moving onto the next stage in your life. Without a doubt, you will continue to gain success in all fields.", "The good luck you have made for yourself is an inspiration to many. May it continue throughout your life.", "May your feet take you…Where your heart wants to go…!!", "So glad to hear about your promotion. May your life always be filled with such good luck.", "Here’s best wishes that your profession brings satisfaction, and pride to you always.", "I am overjoyed to know that your life is filled with such good luck and may it always be that way.", "You can’t start the next chapter of your life if you keep re-reading the last one.", "Happiness is always in your life, may god prosper happiness, wealth and good luck forever in your life.", "Luck is for the lazy..Success is for those who work hard.", "A champion is scared of losing while everyone else is scared of winning.", "There will be no success without hardship. All the best.", "Don’t tell people your dreams. Show them! Best wishes for your future!", "Sending well wishes and plenty of sunshine your way to brighten up your day.", "You are a truly talented person and I wish you well throughout your life.", "May you always get what you want and may you be at the right place. I wish you a very good luck for the future and may you have a wonderful life with no worries.", "Good Luck..Hope you enjoy your new life.", "The best time for new beginnings, is now. All the best.", "Do your best..That’s all you can do!", "May you always walk the glorious road to success. All the best in life. May you achieve greatness and touch all your dreams.", "Your effort is sure to win..Keep up the spirit!", "Set your goals high, and don’t stop till you get there. All the best!", "Close your eyes and make a wish…", "We wanted to wish you well during this very special time in your life.", "Never doubt your abilities for a moment in life..Wish you good luck and ALL THE BEST.", "If you can’t see the bright side of life, polish the dull side.", "Sending best wishes to hope to show that how much I love you.", "From all of us to you, we wish you well and hope that all of your dreams come true.", "Luck is yours, Greetings are mine, You get every success and let ur future be always shine….Best Wishes & Greetings.", "Wish you the very best in everything as you deserve nothing less!", "Wishing your dreams begin to come true, and every tomorrow be happy for you!", "As you start this new journey in life, may every day hold wonderful experiences!", "Be brave to let go of the past. And fight for the present that you deserve. A better future is always yours.", "Here’s wishing you all the best for your endeavors in the coming year!", "Don’t stress. Do your BEST. Forget the rest.", "If you keep working as hard in the future, nothing can stop you. All the best!", "We wish you all the very best for your future life. May everything that you do make you happy and rise.", "Good luck..Hope everything turns out to be simply great for you!", "Only you can control your future.", "Every sunset gives us one day less to live! But every sunrise give us one day more to hope. So, hope for the Best! Good Day!", "If all you can do is crawl, start crawling.", "The secret to a rich life is to have more beginnings than endings.", "Don’t be afraid to move on and start a new chapter.", "Do not wait until the conditions are perfect to begin. Beginning makes the conditions perfect.", "Don’t wait. Make memories today. Celebrate your life!", "Life is a great big canvas, and you should throw all the paint on it you can.", "Life isn’t about finding yourself. Life is about creating yourself. – George Bernard Shaw", "SENDING you GOOD LUCK VIBES.", "Your future will be always be BRIGHT, when you stay focused optimistic and confident.", "Follow your dreams…Never say Never.", "Smile knowing that good luck has finally found you. Wishing you plenty of joy, happiness, and health.", "Don’t be afraid to start over. It’s a brand new opportunity to rebuild what you truly want.", "Just remember – when you think all is lost, the future remains.", "All life is an experiment. The more experiments you make the better.", "I just took a trip to your future in my time machine, and it looks bright. All the best.", "I’m making a prediction that you are going to keep being awesome.", "Good seasons start with good beginnings.", "Take the first step in faith. You don’t have to see the whole staircase, just take the first step.", "You will never win if you never begin.", "Start by doing what’s necessary; then do what’s possible; and suddenly you are doing the impossible.", "Start wherever you are and start small. All the best.", "So happy to hear your wonderful news. Here’s to wishing you the best of luck!", "May all your dreams come true!", "My heart is flowing with joy at knowing that good luck has finally fallen upon your doorstep.", "A brand new start, a new beginning, a new life, make it successful one!", "The next chapter doesn’t start until, you turn the page.", "New beginning, new scopes, new opportunities..Congrats!", "I believe in you! Good luck to you!", "In this new journey of yours, may you get success of tons!", "There is no one more amazing than you. Please accept my best wishes.", "Best wishes that your new degree will be a source of success you so richly deserve.", "Dream it. Wish it. Do it.", "The best way to predict your future is to create it.", "May you follow your dreams and always believe in yourself.", "The secret to living the life of your dreams is to start living the life of your dreams today, in every little way you possibly can.", "May your wishes and dreams come true!", "Please accept our best wishes to you. You are a truly amazing person.", "You are the best author of your future. Best Wishes..!!", "Do what you can, with what you have, where you are. – Theodore Roosevelt", "Because you’ve had a dark past, doesn’t mean you can’t have a bright future.", "Good Luck! Learn from previous ones, excel in the new ones.", "Fly in the plane of ambition; Land in the airport of success. The luck is yours, the wish is mine; May ur future always shine. Good luck.", "Stay positive, work hard, make it happen. All the best.", "Life is a game, play it; Life is a challenge, Meet it; Life is an opportunity, Capture it.", "Shine like starts in the universe. All the best for your future.", "Believe you can and you’re halfway there.", "Wishing you a future full of love and peace..!!", "Nothing we do can change the past, but everything we do changes the future.", "May god shine down his blessings on you today.", "Don’t let your past limit your future. All the best.", "The Future depends on what we do in the PRESENT – Mahatma Gandhi.", "You are an amazing person who deserves all the well wishes in the world.", "The future is as bright as your faith.", "The future belongs to those who believe in the beauty of their dreams.", "The key to happiness is having dreams. The key to success is making them come true. Congratulations.", "Inhale the future. Exhale the past.", "Here’s wishing you good luck for a bright and prosperous future.", "Life can only be understood backwards; but it must be lived forwards.", "May god pour all the luck on you. BEST OF LUCK.", "Dreams and aspirations will be fulfilled everywhere you go. All the best for a bright future.", "I wish you the best of luck for your future. May you get success on every step.", "Good luck in all that comes your way in the future. I hope it goes well for you.", "May all your dreams come true..Yes, ALL of them.", "May you get all the success in your life that you deserve. May all your dreams be fulfilled.", "Wishing you all the very best for your bright future ahead.", "Have firm faith in your strengths and abilities and we wish you all the best for your new responsibilities.", "Hope I will meet you again in my life, Best of luck for your future.", "May God help you in all your endeavors, and always bring success to your doorstep.", "Yesterday was past, today is present and tomorrow is future. May your tomorrow be the best and brightest part of your life. All the best for your future.", "Sending this lucky butterfly to shower you with love, luck & blessings.", "You know that you are the best, leave aside the rest, Go out there and have your say. ALL THE VERY BEST. Do well!", "Remember to keep your head up high and a smile on your face. Wishing you well and all the happiness in the world.", "Wake up and make your dreams come true.", "Best wishes to you. May you experience a lifetime of love, happiness and joy.", "May your path to your destination be easy. Wishing you all the best for your future life.", "May your future be as bright as the rising sun and may get all the happiness in your life. All the best.", "You can’t change the past, so focus on making a great future.", "Change your direction when the world stands against you. Never lose faith from yourself and face all the problems with smile on your face. Best of luck for life and may you have a wonderful life with happiness all the way.", "Sending well wishes to you in the hope that you know how much we love you.", "I wanted to send you a special note letting you know I wish you well.", "Good luck to you and remember to keep your head up high for all the days of your beautiful and wonderful life.", "Life is filled with unexpected adventures. As you continue to move on in life and face new events, adventures, and challenges, we all wish you the best of luck.", "I wish happiness, success, good luck.", "Hope all your plans turn out just right. Good luck!", "Best wishes to you as you progress from one stage in life to another. While this can be a difficult time, the results that you will achieve are worth all of the hard work that you will put in!", "Hope we will stay in touch. Good Luck.", "We dropped by to say we will miss you lots! All the best to you…!", "May you get all the thing you want in life. Best of luck to you my friend.", "Genuine success comes only to those who are ready for it. Good Luck to you.", "I wanted to extend well wishes to you so hold your head up and smile.", "To accomplish great things, we must not only act, but also dream; not only plan, but also believe.", "May good luck smile on you..Always!", "Accept my well wishes to you as you climb the top of every mountain life brings your way.", "If I have to wish you something..I would like to say..All the Best!", "May path you walk on be bright. May success follow you all your life.", "You’re really great and I only want good things to happen to you.", "Wishing you much joy & happiness as you begin a new chapter in your life.", "THE BEST IS YET TO COME.", "I wish you love, peace, and happiness in everything you do as you journey through life.", "I heard about your good news and wanted to wish you good luck and best of health.", "Wish you loads of good luck. Don’t worry and give your best shot, and leave the rest.", "If life is a game. I wish you win always. All the best.", "As you begin a new life together..Here’s me, wishing you the best for your future together!", "You are the one who create your success and happiness. Always maintain happiness, joy and you can face against all difficulties.", "It looks like the future’s really bright. – Micheal P. Anderson", "The future belongs to those who prepare for it today.", "The sweetest things in life are love and happiness. Wishing you good luck now that you have found that.", "May you find as much joy in your life as you have luck my sweet friend.", "I wish you a very happy life ahead in which you do not need to worry about anything.", "I’m confident that you will be blessed. You deserved the best!", "You hold the key to your future. The actions you take every day shape your future.", "May happiness be always there for you in your future. Best wishes for your future.", "I would wish you luck, but I don’t think you need it. You will be successful.", "Even when things are bad, take time to smile. You can still find a way to reach to the top! GOOD LUCK!", "May you get all your desires and may you get success on every single step. I wish you the best of luck.", "Cherish your yesterdays, dream your tomorrows and live your todays.", "I wish you well in all things you do in your life. Best of luck.", "Please accept our well wishes as you move toward a new chapter in your life.", "Sending you happy thoughts and lots of luck as you move forward with your life.", "Life is an open book full of blank pages. You write the story as you go.", "Life is a long journey, forget the journey that have you have already travelled and think about the one that you still have to travel, which is known as future. I wish you all the best for your future.", "Wishing you well in everything you do. You deserve all the best in life.", "A bright future and a whole lot of smiles! Congratulations!"};
+        String[] blessings = {"Dear Lord, Thank You for the woman reading this prayer right now. Thank You for her heart. May You bless her right now. Fill her with Your incredible peace. Wrap her in Your love. May she feel confident and worthy. I pray that she would grow closer to You every day. Fuel a desire deep within her to seek after You. I pray that she would lead a life by the example Christ set. May she face everything with courage and may she walk in integrity. Help her with anything she is struggling with, surround her with encouragement, and give her your precious wisdom. May she experience joy today in Jesus’ name AMEN!", "May success be with you, always…Wishing you good luck!", "May you do well and come out with flying colors…ALL THE BEST!", "May you fly high in life & success be with you always.", "Try a little harder to be a little better. All the best!", "The future is bright and beautiful. Love it, strive for it and work for it!", "My heart is filled with joy at the sound of your good news. Best of luck to you dear friend.", "Be strong as you fight your way to success, the difficulty you face is nothing compared to the joy of success.", "Good luck to you as you move toward the next chapter life brings your way.", "Good luck and tons of best wishes. God bless you in whatever you do. This is my heartiest wish just for you.", "Here’s wishing you SUCCESS in everything you do! Good luck!", "Success comes only to those who believe in themselves and are prepared to win. Good Luck.", "I wish nothing but the best for you.", "Moving on to a new stage in life can be a challenging process. We wish you the best of luck in all of your future endeavors, you will be great!", "It’s kind of fun to do the impossible. All the best.", "Good LUCK to you as you move onto the next stage in your life. Without a doubt, you will continue to have SUCCESS in all areas!", "And here’s wishing you the very best for all the new ventures, that life has in store for you.", "I find my heart singing a chorus at your happy news. Good luck to you during this joyous time.", "Keep your best wishes, close to your heart and watch what happens.", "Don’t be afraid of moving forward. The future is bright.", "If you can dream it, you can do it. Best wishes for your bright future.", "I wish you all the success, happiness, and joy in life.", "You have worked incredibly hard. Good luck! We are proud of you.", "Don’t look back! For a bright future is ahead of you.", "You have proven yourself to be a very skilled individual who has the capacity to do great things with their life. Continue to make us proud as you face new challenges and adventures!", "Dream Big..Sparkle More..Shine Bright.", "Hope your burdens are feeling lighter, and each day is a little bit brighter.", "I am thrilled to know that luck has finally found you. Warmest wishes my friend.", "Sending you thoughtful wishes in all life’s endeavors. Good luck to you now and always.", "Life is what you make of it and I am happy to know that yours is filled with such good luck.", "Wishing you all the best in your future endeavors.", "Winning is not everything but wanting to win is everything Go Ahead…BEST OF LUCK!", "Here’s to wishing you the best of luck in all life has to offer you.", "Go confidently in the direction of your dreams. Live the life you have imagined. – Henry David Thoreau", "Do not go where the path may lead; go instead where there is no path and leave a trail. – Ralph Waldo Emerson", "Hurdles and difficulties are part of the life. Make that part your life a past and think about tomorrow, because what matters at the end of the day is tomorrow.", "Always keep your head up with a smile on the face. Giving best wishes and all the happiness in this world.", "May luck always follow you in whatever you do. May you achieve success as you walk through.", "Changing the face can change nothing, but facing the change can change everything. Good luck.", "Best wishes for you when moving onto the next stage in your life. Without a doubt, you will continue to gain success in all fields.", "The good luck you have made for yourself is an inspiration to many. May it continue throughout your life.", "May your feet take you…Where your heart wants to go…!!", "So glad to hear about your promotion. May your life always be filled with such good luck.", "Here’s best wishes that your profession brings satisfaction, and pride to you always.", "I am overjoyed to know that your life is filled with such good luck and may it always be that way.", "You can’t start the next chapter of your life if you keep re-reading the last one.", "Happiness is always in your life, may god prosper happiness, wealth and good luck forever in your life.", "Luck is for the lazy..Success is for those who work hard.", "A champion is scared of losing while everyone else is scared of winning.", "There will be no success without hardship. All the best.", "Don’t tell people your dreams. Show them! Best wishes for your future!", "Sending well wishes and plenty of sunshine your way to brighten up your day.", "You are a truly talented person and I wish you well throughout your life.", "May you always get what you want and may you be at the right place. I wish you a very good luck for the future and may you have a wonderful life with no worries.", "Good Luck..Hope you enjoy your new life.", "The best time for new beginnings, is now. All the best.", "Do your best..That’s all you can do!", "May you always walk the glorious road to success. All the best in life. May you achieve greatness and touch all your dreams.", "Your effort is sure to win..Keep up the spirit!", "Set your goals high, and don’t stop till you get there. All the best!", "Close your eyes and make a wish…", "We wanted to wish you well during this very special time in your life.", "Never doubt your abilities for a moment in life..Wish you good luck and ALL THE BEST.", "If you can’t see the bright side of life, polish the dull side.", "Sending best wishes to hope to show that how much I love you.", "From all of us to you, we wish you well and hope that all of your dreams come true.", "Luck is yours, Greetings are mine, You get every success and let ur future be always shine….Best Wishes & Greetings.", "Wish you the very best in everything as you deserve nothing less!", "Wishing your dreams begin to come true, and every tomorrow be happy for you!", "As you start this new journey in life, may every day hold wonderful experiences!", "Be brave to let go of the past. And fight for the present that you deserve. A better future is always yours.", "Here’s wishing you all the best for your endeavors in the coming year!", "Don’t stress. Do your BEST. Forget the rest.", "If you keep working as hard in the future, nothing can stop you. All the best!", "We wish you all the very best for your future life. May everything that you do make you happy and rise.", "Good luck..Hope everything turns out to be simply great for you!", "Only you can control your future.", "Every sunset gives us one day less to live! But every sunrise give us one day more to hope. So, hope for the Best! Good Day!", "If all you can do is crawl, start crawling.", "The secret to a rich life is to have more beginnings than endings.", "Don’t be afraid to move on and start a new chapter.", "Do not wait until the conditions are perfect to begin. Beginning makes the conditions perfect.", "Don’t wait. Make memories today. Celebrate your life!", "Life is a great big canvas, and you should throw all the paint on it you can.", "Life isn’t about finding yourself. Life is about creating yourself. – George Bernard Shaw", "SENDING you GOOD LUCK VIBES.", "Your future will be always be BRIGHT, when you stay focused optimistic and confident.", "Follow your dreams…Never say Never.", "Smile knowing that good luck has finally found you. Wishing you plenty of joy, happiness, and health.", "Don’t be afraid to start over. It’s a brand new opportunity to rebuild what you truly want.", "Just remember – when you think all is lost, the future remains.", "All life is an experiment. The more experiments you make the better.", "I just took a trip to your future in my time machine, and it looks bright. All the best.", "I’m making a prediction that you are going to keep being awesome.", "Good seasons start with good beginnings.", "Take the first step in faith. You don’t have to see the whole staircase, just take the first step.", "You will never win if you never begin.", "Start by doing what’s necessary; then do what’s possible; and suddenly you are doing the impossible.", "Start wherever you are and start small. All the best.", "So happy to hear your wonderful news. Here’s to wishing you the best of luck!", "May all your dreams come true!", "My heart is flowing with joy at knowing that good luck has finally fallen upon your doorstep.", "A brand new start, a new beginning, a new life, make it successful one!", "The next chapter doesn’t start until, you turn the page.", "New beginning, new scopes, new opportunities..Congrats!", "I believe in you! Good luck to you!", "In this new journey of yours, may you get success of tons!", "There is no one more amazing than you. Please accept my best wishes.", "Best wishes that your new degree will be a source of success you so richly deserve.", "Dream it. Wish it. Do it.", "The best way to predict your future is to create it.", "May you follow your dreams and always believe in yourself.", "The secret to living the life of your dreams is to start living the life of your dreams today, in every little way you possibly can.", "May your wishes and dreams come true!", "Please accept our best wishes to you. You are a truly amazing person.", "You are the best author of your future. Best Wishes..!!", "Do what you can, with what you have, where you are. – Theodore Roosevelt", "Because you’ve had a dark past, doesn’t mean you can’t have a bright future.", "Good Luck! Learn from previous ones, excel in the new ones.", "Fly in the plane of ambition; Land in the airport of success. The luck is yours, the wish is mine; May ur future always shine. Good luck.", "Stay positive, work hard, make it happen. All the best.", "Life is a game, play it; Life is a challenge, Meet it; Life is an opportunity, Capture it.", "Shine like starts in the universe. All the best for your future.", "Believe you can and you’re halfway there.", "Wishing you a future full of love and peace..!!", "Nothing we do can change the past, but everything we do changes the future.", "May god shine down his blessings on you today.", "Don’t let your past limit your future. All the best.", "The Future depends on what we do in the PRESENT – Mahatma Gandhi.", "You are an amazing person who deserves all the well wishes in the world.", "The future is as bright as your faith.", "The future belongs to those who believe in the beauty of their dreams.", "The key to happiness is having dreams. The key to success is making them come true. Congratulations.", "Inhale the future. Exhale the past.", "Here’s wishing you good luck for a bright and prosperous future.", "Life can only be understood backwards; but it must be lived forwards.", "May god pour all the luck on you. BEST OF LUCK.", "Dreams and aspirations will be fulfilled everywhere you go. All the best for a bright future.", "I wish you the best of luck for your future. May you get success on every step.", "Good luck in all that comes your way in the future. I hope it goes well for you.", "May all your dreams come true..Yes, ALL of them.", "May you get all the success in your life that you deserve. May all your dreams be fulfilled.", "Wishing you all the very best for your bright future ahead.", "Have firm faith in your strengths and abilities and we wish you all the best for your new responsibilities.", "Hope I will meet you again in my life, Best of luck for your future.", "May God help you in all your endeavors, and always bring success to your doorstep.", "Yesterday was past, today is present and tomorrow is future. May your tomorrow be the best and brightest part of your life. All the best for your future.", "Sending this lucky butterfly to shower you with love, luck & blessings.", "You know that you are the best, leave aside the rest, Go out there and have your say. ALL THE VERY BEST. Do well!", "Remember to keep your head up high and a smile on your face. Wishing you well and all the happiness in the world.", "Wake up and make your dreams come true.", "Best wishes to you. May you experience a lifetime of love, happiness and joy.", "May your path to your destination be easy. Wishing you all the best for your future life.", "May your future be as bright as the rising sun and may get all the happiness in your life. All the best.", "You can’t change the past, so focus on making a great future.", "Change your direction when the world stands against you. Never lose faith from yourself and face all the problems with smile on your face. Best of luck for life and may you have a wonderful life with happiness all the way.", "Sending well wishes to you in the hope that you know how much we love you.", "I wanted to send you a special note letting you know I wish you well.", "Good luck to you and remember to keep your head up high for all the days of your beautiful and wonderful life.", "Life is filled with unexpected adventures. As you continue to move on in life and face new events, adventures, and challenges, we all wish you the best of luck.", "I wish happiness, success, good luck.", "Hope all your plans turn out just right. Good luck!", "Best wishes to you as you progress from one stage in life to another. While this can be a difficult time, the results that you will achieve are worth all of the hard work that you will put in!", "Hope we will stay in touch. Good Luck.", "We dropped by to say we will miss you lots! All the best to you…!", "May you get all the thing you want in life. Best of luck to you my friend.", "Genuine success comes only to those who are ready for it. Good Luck to you.", "I wanted to extend well wishes to you so hold your head up and smile.", "To accomplish great things, we must not only act, but also dream; not only plan, but also believe.", "May good luck smile on you..Always!", "Accept my well wishes to you as you climb the top of every mountain life brings your way.", "If I have to wish you something..I would like to say..All the Best!", "May path you walk on be bright. May success follow you all your life.", "You’re really great and I only want good things to happen to you.", "Wishing you much joy & happiness as you begin a new chapter in your life.", "THE BEST IS YET TO COME.", "I wish you love, peace, and happiness in everything you do as you journey through life.", "I heard about your good news and wanted to wish you good luck and best of health.", "Wish you loads of good luck. Don’t worry and give your best shot, and leave the rest.", "If life is a game. I wish you win always. All the best.", "As you begin a new life together..Here’s me, wishing you the best for your future together!", "You are the one who create your success and happiness. Always maintain happiness, joy and you can face against all difficulties.", "It looks like the future’s really bright. – Micheal P. Anderson", "The future belongs to those who prepare for it today.", "The sweetest things in life are love and happiness. Wishing you good luck now that you have found that.", "May you find as much joy in your life as you have luck my sweet friend.", "I wish you a very happy life ahead in which you do not need to worry about anything.", "I’m confident that you will be blessed. You deserved the best!", "You hold the key to your future. The actions you take every day shape your future.", "May happiness be always there for you in your future. Best wishes for your future.", "I would wish you luck, but I don’t think you need it. You will be successful.", "Even when things are bad, take time to smile. You can still find a way to reach to the top! GOOD LUCK!", "May you get all your desires and may you get success on every single step. I wish you the best of luck.", "Cherish your yesterdays, dream your tomorrows and live your todays.", "I wish you well in all things you do in your life. Best of luck.", "Please accept our well wishes as you move toward a new chapter in your life.", "Sending you happy thoughts and lots of luck as you move forward with your life.", "Life is an open book full of blank pages. You write the story as you go.", "Life is a long journey, forget the journey that have you have already travelled and think about the one that you still have to travel, which is known as future. I wish you all the best for your future.", "Wishing you well in everything you do. You deserve all the best in life.", "A bright future and a whole lot of smiles! Congratulations!"};
 
         return blessings;
     }
